+++ src/main/java/protocol/request/AdminDeleteUserRequest.java
@@ -19,7 +19,7 @@ public class AdminDeleteUserRequest extends Request<AdminDeleteUserRequest.Paylo
 
     public record Payload(
             @Positive
-            int registro
+            long registro
     ) {
     }
 }
