+++ src/main/java/server/dto/UpdateUser.java
@@ -3,5 +3,5 @@ package server.dto;
 import lombok.Builder;
 
 @Builder
-public record UpdateUser(int registro, String email, String nome, String senha, Boolean tipo) {
+public record UpdateUser(long registro, String email, String nome, String senha, Boolean tipo) {
 }
